Add method controller (#2)
* Add few features

* Change to sdev-grpcui

* cleanup old jquery files

* Add json Editor support

* Add method controller name

* cleanup

---------
