+++ internal/resources/webform/webform-template.html
@@ -31,6 +31,7 @@
       </td>
       <td id="grpc-descriptions">
         <button id="grpc-descriptions-toggle">»</button>
+        <pre id="grpc-controller-description"></pre>
         <pre id="grpc-service-description"></pre>
         <pre id="grpc-method-description"></pre>
         <pre id="grpc-service-description-end"></pre>
